move the role permissions constants
