+++ flashcard.wars/build.gradle
@@ -29,7 +29,7 @@ dependencies {
 	implementation 'org.apache.commons:commons-lang3:3.0'
 	implementation 'com.google.guava:guava:33.2.0-jre'
 	implementation 'com.sun.mail:javax.mail:1.6.2'
-	compileOnly 'org.mapstruct:mapstruct-processor:1.6.2'
+	implementation 'org.mapstruct:mapstruct-processor:1.6.2'
 	compileOnly 'org.projectlombok:lombok'
 	annotationProcessor 'org.projectlombok:lombok'
 	runtimeOnly 'org.postgresql:postgresql'
