+++ flashcard.wars/src/main/java/com/games/flashcard/util/RolePermissions.java
@@ -1,4 +1,4 @@
-package com.games.flashcard;
+package com.games.flashcard.util;
 
 import com.games.flashcard.model.enums.PERMISSION;
